+++ WooCommerce/Classes/ServiceLocator/ServiceLocator.swift
@@ -88,7 +88,7 @@ final class ServiceLocator {
     ///
     static var shippingSettingsService: ShippingSettingsService {
         guard let shippingSettingsService = _shippingSettingsService else {
-            let siteID = stores.sessionManager.defaultStoreID ?? Int.min
+            let siteID = stores.sessionManager.defaultStoreID ?? Int64.min
             let service = StorageShippingSettingsService(siteID: Int64(siteID),
                                                          storageManager: storageManager)
             _shippingSettingsService = service
